Test: compute at-inferred within a non-top-level context
The at-inferred macro, when presented with a kwargs call, previously would ask return types about the function directly. This is an accurate representation of how the call behaves at top/global scope, but typically we are much more interested in the behavior of the function when called from another function with well-typed arguments. Within this context, the compiler can use inlining and other optimizations that allow some keyword functions to recover type-stability. This is becoming a common idiom within base.

This patch changes the behavior of at-inferred to test the behavior of the `f(args..., kwargs...)` expression as though it were placed directly as written within an anonymous function.  Required for #25831.

(cherry picked from commit 7cf358606b4c26cd953c8551c0882ad24e57a300)